+++ firebase-vertexai/CHANGELOG.md
@@ -1,5 +1,5 @@
 # Unreleased
-
+* [changed] Made `FunctionCallPart.args` nullable.
 
 # 16.0.0-beta03
 * [changed] Breaking Change: changed `Schema.int` to return 32 bit integers instead of 64 bit (long).